Simplify ccwrap
It seems that all ccwrap really does is add the compiler-specific include
directories, which -nostdinc prevents being added to the include path.

Work these out once, rather than inside this wrapper.

(As written this isn't strictly correct as we should split these headers
into two groups, used for C and C++, and only use the C ones in when
invoking the C compiler, but both for the C++ compiler.)

Also stop using ccwrap on MinGW compiled utilities.

The one fly in the ointment is dumper.exe, which is a cygwin executable,
linked with cygwin's libbfd. We should learn where bfd.h is in the same
fashion as windef.h

Future work: It should now be possible to transform all the include
directories passed into ccwrap into simple -I or -isystem flags, remove
ccwrap, and use the compiler directly.